refactor - split  tables into components and update state and queries
