+++ packages/sections/src/drug/ClinicalIndications/Body.tsx
@@ -1,47 +1,29 @@
 import { useQuery } from "@apollo/client";
 
 import localData from "./clinical_indication_CHEMBL2105708.json"  // !! IMPORT LOCAL DATA FOR NOW !!
-
-import { Link, SectionItem, PaginationActionsComplete, TableDrawer, OtTable } from "ui";
+import { Grid } from "@mui/material";
+import { SectionItem, PaginationActionsComplete, OtTable } from "ui";
 import { sourceMap, phaseMap } from "@ot/constants";
 import { referenceUrls } from "@ot/utils";
 
+import { useState } from "react";
 import Description from "./Description";
-import RecordsDrawer from "./RecordsDrawer";
-import SideBySideTables from "./SideBySideTables";
 import CLINICAL_INDICATIONS_QUERY from "./ClinicalIndicationsQuery.gql";
 import { definition } from ".";
-
-const columns = [
-  {
-    id: "diseaseId",
-    label: "Indication",
-    renderCell: ({ diseaseName, diseaseId }) => (
-      <Link asyncTooltip to={`/disease/${diseaseId}`}>
-        {diseaseName}
-      </Link>
-    ),
-  },
-  {
-    id: "maxClinicalStatus",
-    label: "Max clinical status",
-  },
-  {
-    id: "mappingConfidence",
-    label: "Mapping confidence"
-  },
-  {
-    id: "records",
-    label: "Approvals and Trials",
-    renderCell: ({ clinicalReportIds }) => (
-      <RecordsDrawer recordIds={clinicalReportIds} />
-    ),
-  },
-];
+import IndicationsTable from "./IndicationsTable";
+import RecordsTable from "./RecordsTable";
 
 function Body({ id: chemblId, label: name, entity }) {
   const variables = { chemblId };
   const request = useQuery(CLINICAL_INDICATIONS_QUERY, { variables });
+
+  // disease of selected row in first table
+  const [selectedDisease, setSelectedDisease] = useState("");
+  
+  // records to be showing in 2nd table
+  const [records, setRecords] = useState([]);
+
+  const rows = localData;  // !! CHECNGE TO ROWS FROM REQUEST WHEN HAVE API !!
 
   return (
     <SectionItem
@@ -51,29 +33,28 @@ function Body({ id: chemblId, label: name, entity }) {
       renderDescription={() => <Description name={name} />}
       renderBody={() => {
         return (
-          <>
-            <OtTable
-              columns={columns}
-              dataDownloader
-              dataDownloaderFileStem={`${chemblId}-indications`}
-              rows={localData}
-              // rows={request.data?.drug.indications.rows}
-              showGlobalFilter
-              // sortBy="maxPhaseForIndication"
-              // order="desc"
-              // ActionsComponent={PaginationActionsComplete}
-              query={CLINICAL_INDICATIONS_QUERY.loc.source.body}
-              variables={variables}
-              loading={request.loading}
-            />
-            <div style={{ height: "50px"}}></div>
-            <SideBySideTables
-              rows={localData}
-              query={CLINICAL_INDICATIONS_QUERY.loc.source.body}
-              variables={variables}
-              loading={request.loading}
-            />
-          </>
+          <Grid container spacing={10}>
+            {/* LHS table */}
+            <Grid item xs={12} md={6}>
+              {rows?.length > 0 && (  
+                <IndicationsTable
+                  rows={rows}
+                  setSelectedDisease={setSelectedDisease}
+                  setRecords={setRecords}
+                  query={CLINICAL_INDICATIONS_QUERY.loc.source.body}
+                  variables={variables}
+                  loading={request.loading}
+                />
+              )}
+            </Grid>
+
+            {/* RHS table */}
+            <Grid item xs={12} md={6}>
+              {records && (
+                <RecordsTable selectedDisease={selectedDisease} records={records}/>
+              )}
+            </Grid>
+          </Grid>
         );
       }}
     />
