+++ packages/sections/src/drug/ClinicalIndications/ClinicalRecordsQuery.gql
@@ -1,0 +1,5 @@
+query ClinicalRecordsQuery($chemblId: String!) {
+  drug(chemblId: $chemblId) {
+    id  # !! UPDATE ONCE HAVE API !!
+  }
+}
